Add basic CMake support
This allows for better integration in other projects which are already
using CMake.

If no option are provided a static and a shared library, as well as
the CLI are going to be built.

Building of libraries can be disabled with the corresponding option, as
well as the CLI. Note that the CLI depends on the static library.

-Wall -Wextra are enabled by default.